feat: Add fractional shares support for crypto and mutual funds
- Update Transaction and Holding classes to support float shares instead of int
- Add supports_fractional_shares property to AssetProvider and Asset classes
- Correct logic: only crypto (BTC-USD, ETH-USD) and mutual funds (VF*, VG*) support fractional shares
- ETFs like BIL and SPY do NOT support fractional shares (corrected from previous implementation)
- Update backtest functions to accept float initial_qty instead of int
- Enable dividend reinvestment to create fractional shares
- Maintain backward compatibility for integer share operations

This enables proper handling of crypto fractional shares, mutual fund fractional shares, and dividend reinvestment scenarios.

+++ src/compare/batch_comparison.py
@@ -46,7 +46,7 @@ def run_single_backtest(
     start_date: date,
     end_date: date,
     algo_name: str,
-    initial_qty: int = 10000,
+    initial_qty: float = 10000,
     reference_rate_ticker: str = "",
     risk_free_rate_ticker: str = "",
 ) -> Dict[str, Any]:
@@ -203,7 +203,7 @@ def run_batch_comparison(
     ticker: str,
     start_date: date,
     end_date: date,
-    initial_qty: int = 10000,
+    initial_qty: float = 10000,
     configs: Optional[List[str]] = None,
     reference_asset: str = "VOO",
     risk_free_asset: str = "BIL",
